@@ -5,13 +5,13 @@ author: Emanuel Braz - fastencoding@gmail.com
 homepage: https://github.com/emanuel-braz/one_context.git
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  shared_preferences: ^0.5.7+3
-  back_button_interceptor: ^4.4.0
+  shared_preferences: ^2.0.5
+  back_button_interceptor: ^5.0.0
 
 dev_dependencies:
   flutter_test:
