@@ -4,9 +4,9 @@ import 'package:one_context/src/components/one_basic_widget.dart';
 import 'package:one_context/src/controllers/one_context.dart';
 
 class OneContextWidget extends StatefulWidget {
-  final Widget child;
+  final Widget? child;
 
-  OneContextWidget({Key key, this.child}) : super(key: key);
+  OneContextWidget({Key? key, this.child}) : super(key: key);
 
   _OneContextWidgetState createState() => _OneContextWidgetState();
 }
@@ -16,11 +16,11 @@ class _OneContextWidgetState extends State<OneContextWidget> {
   void initState() {
     super.initState();
     OneContext().registerDialogCallback(
-        showDialog: _showDialog,
-        showSnackBar: _showSnackBar,
-        showModalBottomSheet: _showModalBottomSheet,
-        showBottomSheet: _showBottomSheet,
-        showDatePicker: _showDatePicker);
+      showDialog: _showDialog,
+      showSnackBar: _showSnackBar,
+      showModalBottomSheet: _showModalBottomSheet,
+      showBottomSheet: _showBottomSheet,
+    );
     BackButtonInterceptor.add(backButtonInterceptor);
   }
 
@@ -34,13 +34,13 @@ class _OneContextWidgetState extends State<OneContextWidget> {
     try {
       if (OneContext().hasDialogVisible) {
         OneBasicWidget lastDialog = OneContext().dialogList.last;
-        if (lastDialog.isBackButtonDismissible) {
+        if (lastDialog.isBackButtonDismissible ?? true) {
           if (lastDialog.type == OneBasicWidgetTypes.snackbar) {
             OneContext().hideCurrentSnackBar();
           } else {
             OneContext().popDialog();
           }
-          lastDialog?.onClickBackButtonDismissCallback?.call();
+          lastDialog.onClickBackButtonDismissCallback?.call();
         }
         return true;
       } else
@@ -57,98 +57,54 @@ class _OneContextWidgetState extends State<OneContextWidget> {
       body: Builder(
         builder: (innerContext) {
           OneContext().context = innerContext;
-          return widget.child;
+          return widget.child!;
         },
       ),
     );
   }
 
-  Future<T> _showDialog<T>({
-    bool barrierDismissible = true,
-    Widget Function(BuildContext) builder,
+  Future<T?> _showDialog<T>({
+    bool? barrierDismissible = true,
+    required Widget Function(BuildContext) builder,
     bool useRootNavigator = true,
     Color barrierColor = Colors.transparent,
   }) =>
-      showDialog<T>(
+      showDialog<T?>(
         context: context,
         builder: (context) => builder(context),
-        barrierDismissible: barrierDismissible,
+        barrierDismissible: barrierDismissible!,
         useRootNavigator: useRootNavigator,
         barrierColor: barrierColor,
       );
 
-  ScaffoldFeatureController<SnackBar, SnackBarClosedReason> _showSnackBar(SnackBar Function(BuildContext) builder) =>
-      Scaffold.of(OneContext().context).showSnackBar(builder(OneContext().context));
+  ScaffoldFeatureController<SnackBar, SnackBarClosedReason> _showSnackBar(SnackBar Function(BuildContext?) builder) =>
+      Scaffold.of(OneContext().context!).showSnackBar(builder(OneContext().context));
 
-  Future<T> _showModalBottomSheet<T>(
-      {Widget Function(BuildContext) builder,
-      Color backgroundColor,
-      double elevation,
-      ShapeBorder shape,
-      Clip clipBehavior,
-      bool isScrollControlled = false,
-      bool useRootNavigator = false,
-      bool isDismissible = true}) {
+  Future<T?> _showModalBottomSheet<T>(
+      {required Widget Function(BuildContext) builder,
+      Color? backgroundColor,
+      double? elevation,
+      ShapeBorder? shape,
+      Clip? clipBehavior,
+      bool? isScrollControlled = false,
+      bool? useRootNavigator = false,
+      bool? isDismissible = true}) {
     return showModalBottomSheet<T>(
       context: context,
       builder: builder,
       backgroundColor: backgroundColor,
       clipBehavior: clipBehavior,
       elevation: elevation,
-      isDismissible: isDismissible,
-      isScrollControlled: isScrollControlled,
+      isDismissible: isDismissible!,
+      isScrollControlled: isScrollControlled!,
       shape: shape,
-      useRootNavigator: useRootNavigator,
+      useRootNavigator: useRootNavigator!,
     );
   }
 
   PersistentBottomSheetController<T> _showBottomSheet<T>(
-      {@required Widget Function(BuildContext) builder, Color backgroundColor, double elevation, ShapeBorder shape, Clip clipBehavior}) {
+      {Widget Function(BuildContext)? builder, Color? backgroundColor, double? elevation, ShapeBorder? shape, Clip? clipBehavior}) {
     return showBottomSheet<T>(
-        context: OneContext().context, builder: builder, backgroundColor: backgroundColor, elevation: elevation, shape: shape, clipBehavior: clipBehavior);
+        context: OneContext().context!, builder: builder!, backgroundColor: backgroundColor, elevation: elevation, shape: shape, clipBehavior: clipBehavior);
   }
-
-  Future<DateTime> _showDatePicker({
-    @required DateTime initialDate,
-    @required DateTime firstDate,
-    @required DateTime lastDate,
-    DateTime currentDate,
-    DatePickerEntryMode initialEntryMode,
-    SelectableDayPredicate selectableDayPredicate,
-    String helpText,
-    String cancelText,
-    String confirmText,
-    Locale locale,
-    bool useRootNavigator,
-    RouteSettings routeSettings,
-    TextDirection textDirection,
-    TransitionBuilder builder,
-    DatePickerMode initialDatePickerMode,
-    String errorFormatText,
-    String errorInvalidText,
-    String fieldHintText,
-    String fieldLabelText,
-  }) =>
-      showDatePicker(
-        context: OneContext().context,
-        initialDate: initialDate,
-        firstDate: firstDate,
-        lastDate: lastDate,
-        currentDate: currentDate,
-        initialEntryMode: initialEntryMode,
-        selectableDayPredicate: selectableDayPredicate,
-        helpText: helpText,
-        cancelText: cancelText,
-        confirmText: confirmText,
-        locale: locale,
-        useRootNavigator: useRootNavigator,
-        routeSettings: routeSettings,
-        textDirection: textDirection,
-        builder: builder,
-        initialDatePickerMode: initialDatePickerMode,
-        errorFormatText: errorFormatText,
-        errorInvalidText: errorInvalidText,
-        fieldHintText: fieldHintText,
-        fieldLabelText: fieldLabelText,
-      );
 }
